fix: move cw-multi test
