+++ Cargo.toml
@@ -14,10 +14,11 @@ version="0.1.0"
 
 [workspace.dependencies]
 cosmwasm-schema = {version="1.4.0",default-features = false}
-cosmwasm-std = {version="1.4.0",default-features = false,features = ["ibc3"]}
+cosmwasm-std = {version="1.4.0",default-features = false}
 schemars = {version="0.8.10",default-features = false}
 serde = { version = "1.0.145", default-features = false, features = ["derive"] }
 thiserror = { version = "1.0.31" }
+
 cw-storage-plus = {git="https://github.com/icon-project/cw-storage-plus.git", branch="fix-raw"}
 
 cw-xcall-lib={package="cw-xcall-lib", git="https://github.com/icon-project/xCall.git", branch="fix/cosmwasm-1.3",default-features = false}
