+++ libraries/rust/test-utils/Cargo.toml
@@ -14,8 +14,8 @@ ibc-proto = { workspace=true}
 hex={workspace=true}
 hex-literal={workspace=true}
 common={workspace=true}
-cw-multi-test = "0.15.1"
+
 
 [dev-dependencies]
-
+cw-multi-test = "0.15.1"
 getrandom = {version = "0.2", default-features = false, features = ["custom"]}
