+++ libraries/rust/test-utils/src/lib.rs
@@ -6,15 +6,13 @@ use std::{
     path::PathBuf,
 };
 
-use common::ibc::events::IbcEventType;
-use cw_multi_test::AppResponse;
 use ibc_proto::ibc::core::channel::v1::Packet;
 use serde::Deserialize;
 
 use common::icon::icon::types::v1::BtpHeader;
 use common::icon::icon::types::v1::MerkleNode;
 use common::icon::icon::types::v1::SignedHeader;
-use cosmwasm_std::{Attribute, Event};
+use cosmwasm_std::Attribute;
 
 #[derive(Debug, Deserialize, Default, Clone)]
 pub struct IntegrationData {
@@ -255,21 +253,4 @@ pub fn to_attribute_map(attrs: &Vec<Attribute>) -> HashMap<String, String> {
         map.insert(attr.key.clone(), attr.value.clone());
     }
     map
-}
-
-pub fn get_event(res: &AppResponse, event: &str) -> Option<HashMap<String, String>> {
-    let event = res
-        .events
-        .iter()
-        .filter(|e| e.ty == event)
-        .collect::<Vec<&Event>>();
-    if !event.is_empty() {
-        let map = to_attribute_map(&event[0].attributes);
-        return Some(map);
-    }
-    None
-}
-
-pub fn get_event_name(event_type: IbcEventType) -> String {
-    format!("wasm-{}", event_type.as_str())
 }
